requirements.txt
requests~=2.32.3
aiohttp~=3.10.10
pillow~=11.0.0
aiogram~=3.14.0
python-dotenv~=1.0.1